refactor(sdk-migration): reorganize packages into client and tools subpackages
