+++ pkg/mcp-official/tools/tools.go
@@ -1,17 +1,18 @@
-package mcpofficial
+package tools
 
 import (
 	"github.com/hashicorp/terraform-mcp-server/pkg/toolsets"
 	"github.com/modelcontextprotocol/go-sdk/mcp"
 	log "github.com/sirupsen/logrus"
+	tfeTools "github.com/hashicorp/terraform-mcp-server/pkg/mcp-official/tools/tfe"
 )
 
 func RegisterTools(svr *mcp.Server, logger *log.Logger, enabledToolsets []string) {
 	if toolsets.IsToolEnabled("list_workspaces", enabledToolsets) {
-		mcp.AddTool(svr, ListWorkpsacesTool(), ListWorkspacesFunc)
+		mcp.AddTool(svr, tfeTools.ListWorkpsacesTool(), tfeTools.ListWorkspacesFunc)
 	}
 
 	if toolsets.IsToolEnabled("list_terraform_orgs", enabledToolsets) {
-		mcp.AddTool(svr, ListTerraformOrganizationsTool(), ListTerraformOrganizationsFunc)
+		mcp.AddTool(svr, tfeTools.ListTerraformOrganizationsTool(), tfeTools.ListTerraformOrganizationsFunc)
 	}
 }
